Split out Image context in parser
