+++ fonduer/parser/models/__init__.py
@@ -3,18 +3,17 @@ from fonduer.parser.models.context import (
     Context,
     Document,
     Figure,
-    Image,
     ImplicitSpan,
     Phrase,
     Span,
     Table,
-    TemporaryImage,
     TemporaryImplicitSpan,
     TemporarySpan,
     Webpage,
     construct_stable_id,
     split_stable_id,
 )
+from fonduer.parser.models.image import Image, TemporaryImage
 
 __all__ = [
     "Cell",
